Rearrange Code: Move definition of URLMetadataModel to url_metadata.py
