+++ datalad_registry/blueprints/api/url_metadata.py
@@ -2,9 +2,8 @@
 # i.e. the metadata of datasets at individual URLs.
 
 from flask_openapi3 import Tag
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, Field, StrictStr
 
-from datalad_registry.com_models import URLMetadataModel
 from datalad_registry.models import URLMetadata, db
 
 from . import bp
@@ -20,6 +19,23 @@ class _PathParams(BaseModel):
     """
 
     url_metadata_id: int = Field(..., description="The ID of the URL metadata")
+
+
+class URLMetadataModel(BaseModel):
+    """
+    Pydantic model for representing the database model URLMetadata for communication
+    in JSON
+    """
+
+    dataset_describe: StrictStr
+    dataset_version: StrictStr
+    extractor_name: StrictStr
+    extractor_version: StrictStr
+    extraction_parameter: dict
+    extracted_metadata: dict
+
+    class Config:
+        orm_mode = True
 
 
 @bp.get(
