+++ datalad_registry/com_models.py
@@ -20,20 +20,3 @@ class MetaExtractResult(BaseModel):
     action: Literal["meta_extract"]
     status: StrictStr
     metadata_record: MetadataRecord
-
-
-class URLMetadataModel(BaseModel):
-    """
-    Pydantic model for representing the database model URLMetadata for communication
-    in JSON
-    """
-
-    dataset_describe: StrictStr
-    dataset_version: StrictStr
-    extractor_name: StrictStr
-    extractor_version: StrictStr
-    extraction_parameter: dict
-    extracted_metadata: dict
-
-    class Config:
-        orm_mode = True
