+++ datalad_registry/tests/test_new_organization/test_blueprints/test_api/test_url_metadata.py
@@ -1,6 +1,6 @@
 import pytest
 
-from datalad_registry.com_models import URLMetadataModel
+from datalad_registry.blueprints.api.url_metadata import URLMetadataModel
 
 
 @pytest.fixture
